refactor(preview): decompose PreviewPanel into hooks and toolbar
PreviewPanel.tsx: 659 lines -> 116 lines
Extracted:
- hooks/usePreviewBridge.ts: IPC sync, navigation events, ResizeObserver
- hooks/usePreviewCapture.ts: capture handlers and busy states
- PreviewToolbar.tsx: presentational navigation and capture buttons
