+++ apps/web/src/components/panels/hooks/usePreviewBridge.ts
@@ -1,0 +1,256 @@
+import {
+  useCallback,
+  useEffect,
+  useRef,
+  useState,
+  type RefObject,
+} from "react";
+import { useDiffStore } from "@/stores/diffStore";
+
+const NAV_ERROR_LABEL: Record<string, string> = {
+  "no-bounds": "Wait for the panel to finish layout, then try again.",
+  "invalid-url": "Only http and https URLs are allowed.",
+  "empty-url": "Enter a URL.",
+  "no-window": "Preview is unavailable.",
+};
+
+/** Resolves an IPC error code to a short user-visible hint. */
+export function formatNavError(code: string): string {
+  return NAV_ERROR_LABEL[code] ?? code;
+}
+
+export interface UsePreviewBridgeOptions {
+  readonly threadId: string;
+  readonly workspaceId?: string | null;
+  readonly surfaceRef: RefObject<HTMLDivElement | null>;
+}
+
+export interface PreviewBridgeState {
+  readonly inputUrl: string;
+  readonly setInputUrl: (url: string) => void;
+  readonly navError: string | null;
+  readonly setNavError: (err: string | null) => void;
+  readonly canBack: boolean;
+  readonly canFwd: boolean;
+  readonly previewLoading: boolean;
+  readonly pageTitle: string | null;
+  readonly faviconUrl: string | null;
+  readonly storedUrl: string;
+  readonly pushSync: (visible: boolean) => Promise<void>;
+  readonly refreshNav: () => Promise<void>;
+  readonly onGoBack: () => Promise<void>;
+  readonly onGoForward: () => Promise<void>;
+  readonly onReload: () => Promise<void>;
+  readonly onOpenExternal: () => Promise<void>;
+  readonly onNavigate: (url: string) => void;
+}
+
+/**
+ * Manages IPC connection to the Electron preview: bounds sync, navigation
+ * state, ResizeObserver tracking, and navigation event listeners.
+ */
+export function usePreviewBridge({
+  threadId,
+  workspaceId,
+  surfaceRef,
+}: UsePreviewBridgeOptions): PreviewBridgeState {
+  const [inputUrl, setInputUrl] = useState("");
+  const [navError, setNavError] = useState<string | null>(null);
+  const [canBack, setCanBack] = useState(false);
+  const [canFwd, setCanFwd] = useState(false);
+  const [previewLoading, setPreviewLoading] = useState(false);
+  const [pageTitle, setPageTitle] = useState<string | null>(null);
+  const [faviconUrl, setFaviconUrl] = useState<string | null>(null);
+
+  const storedUrl = useDiffStore(
+    (s) => s.previewUrlByThread[threadId] ?? "",
+  );
+
+  /** Stable ref to the current storedUrl, read inside pushSync to avoid
+   *  adding storedUrl to pushSync's dependency array. */
+  const storedUrlRef = useRef(storedUrl);
+  storedUrlRef.current = storedUrl;
+
+  useEffect(() => {
+    setInputUrl(storedUrl);
+    setPageTitle(null);
+    setFaviconUrl(null);
+    setNavError(null);
+  }, [threadId, storedUrl]);
+
+  const refreshNav = useCallback(async () => {
+    const preview = window.desktopBridge?.preview;
+    if (!preview) return;
+    const s = await preview.getNavigationState();
+    setCanBack(s.canGoBack);
+    setCanFwd(s.canGoForward);
+  }, []);
+
+  const pushSync = useCallback(
+    async (visible: boolean) => {
+      const preview = window.desktopBridge?.preview;
+      if (!preview) return;
+      const el = surfaceRef.current;
+      const hint = storedUrlRef.current.trim() || null;
+      if (!visible || !el) {
+        await preview.sync({
+          visible: false,
+          bounds: null,
+          threadId,
+          resumeUrlHint: hint,
+          workspaceId: workspaceId ?? null,
+        });
+        return;
+      }
+      const r = el.getBoundingClientRect();
+      await preview.sync({
+        visible: true,
+        bounds: {
+          x: Math.round(r.left),
+          y: Math.round(r.top),
+          width: Math.round(r.width),
+          height: Math.round(r.height),
+        },
+        threadId,
+        resumeUrlHint: hint,
+        workspaceId: workspaceId ?? null,
+      });
+    },
+    [threadId, workspaceId, surfaceRef],
+  );
+
+  const pushSyncRef = useRef(pushSync);
+  pushSyncRef.current = pushSync;
+  const refreshNavRef = useRef(refreshNav);
+  refreshNavRef.current = refreshNav;
+
+  useEffect(() => {
+    const preview = window.desktopBridge?.preview;
+    if (!preview) return;
+    const unsub = preview.onDidNavigate((p) => {
+      if (
+        p.url &&
+        !p.url.startsWith("chrome-error://") &&
+        !p.url.startsWith("about:")
+      ) {
+        useDiffStore.getState().setPreviewUrlForThread(threadId, p.url);
+        setInputUrl(p.url);
+        setPageTitle(p.title ?? null);
+        setFaviconUrl(p.favicon ?? null);
+      } else {
+        setPageTitle(null);
+        setFaviconUrl(null);
+      }
+      void refreshNav();
+    });
+    return unsub;
+  }, [threadId, refreshNav]);
+
+  useEffect(() => {
+    const preview = window.desktopBridge?.preview;
+    if (!preview?.onDidUpdateFavicon) return;
+    return preview.onDidUpdateFavicon((p) => {
+      setFaviconUrl(p.favicon);
+    });
+  }, []);
+
+  useEffect(() => {
+    const preview = window.desktopBridge?.preview;
+    if (!preview) return;
+    return preview.onLoadingState((p) => setPreviewLoading(p.loading));
+  }, []);
+
+  useEffect(() => {
+    const preview = window.desktopBridge?.preview;
+    if (!preview) return;
+    const el = surfaceRef.current;
+    if (!el) return;
+
+    let mounted = true;
+    let raf = 0;
+    const schedule = () => {
+      if (!mounted) return;
+      if (raf) cancelAnimationFrame(raf);
+      raf = requestAnimationFrame(() => {
+        raf = 0;
+        if (!mounted) return;
+        void pushSyncRef.current(true);
+        void refreshNavRef.current();
+      });
+    };
+
+    const ro = new ResizeObserver(schedule);
+    ro.observe(el);
+    schedule();
+
+    window.addEventListener("resize", schedule);
+    return () => {
+      mounted = false;
+      ro.disconnect();
+      window.removeEventListener("resize", schedule);
+      if (raf) cancelAnimationFrame(raf);
+      void pushSyncRef.current(false);
+    };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  const onGoBack = useCallback(async () => {
+    const preview = window.desktopBridge?.preview;
+    if (!preview) return;
+    await pushSync(true);
+    await preview.goBack();
+    await refreshNav();
+  }, [pushSync, refreshNav]);
+
+  const onGoForward = useCallback(async () => {
+    const preview = window.desktopBridge?.preview;
+    if (!preview) return;
+    await pushSync(true);
+    await preview.goForward();
+    await refreshNav();
+  }, [pushSync, refreshNav]);
+
+  const onReload = useCallback(async () => {
+    const preview = window.desktopBridge?.preview;
+    if (!preview) return;
+    await pushSync(true);
+    await preview.reload();
+    await refreshNav();
+  }, [pushSync, refreshNav]);
+
+  const onOpenExternal = useCallback(async () => {
+    const preview = window.desktopBridge?.preview;
+    if (!preview) return;
+    await preview.openExternal();
+  }, []);
+
+  const onNavigate = useCallback(
+    (url: string) => {
+      setInputUrl(url);
+      void window.desktopBridge?.preview.navigate(url).then((r) => {
+        if (!r.ok) setNavError(formatNavError(r.error));
+      });
+    },
+    [],
+  );
+
+  return {
+    inputUrl,
+    setInputUrl,
+    navError,
+    setNavError,
+    canBack,
+    canFwd,
+    previewLoading,
+    pageTitle,
+    faviconUrl,
+    storedUrl,
+    pushSync,
+    refreshNav,
+    onGoBack,
+    onGoForward,
+    onReload,
+    onOpenExternal,
+    onNavigate,
+  };
+}
